Fix tests
It appears that due to the pachyderm histogram API change, small changes
were made to the numerical values of a number of arrays. To address this
issue, the tolerances were reduced from a relative tolerance to an
absolute tolerance. All absolute tolerances are below 1e-4.

Further, the image comparison was made more exact in an effort to make
the tests a bit more stringent