Add possibility to calculate minos errors
They aren't stored because they can't be used to obtain a covariance
matrix. But they can be compared with the Hesse errors. If they are
similar, that means that the function at the minima is well approximated
by a hyperparabola, which means that we are okay to use the Hesse
errors.