Update signal inclusive baseline comparison image
Since the plotting style changed (adding the background function to
signal plots and moving the legend to the signal inclusive axis), the
baseline image needs to be updated. Note that the underlying fit code
(ie the tests without the image comparison) still agree perfectly, so
all that we're changing is the baseline image.