Add correlation matrix to fit results
It's calculated from the covariance matrix. The calculation is checked
against minuit (from which the calculation is derived), and is tested to
work properly with YAML (which was not entirely obvious because of some
interference caused by unknowingly returning a `np.float64` which isn't
known to YAML. Note that this presented in a very frustrating way, saying
only that `1.0` couldn't be represented. The type wasn't obvious because
the repr doesn't specify the (numpy) type).